requirements.txt
fastapi==0.109.2
pydantic==2.6.4
pydantic_core==2.16.3
PyMySQL==1.1.0
python-decouple==3.8
requests==2.31.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.27.1
Pillow==10.4.0
pytesseract==0.3.10
PyJWT==2.8.0 
aiohttp==3.9.5 
aiohttp-retry==2.8.3
aiosignal==1.3.1 
attrs==23.2.0 
frozenlist==1.4.1 
multidict==6.0.5 
twilio==9.2.3
yarl==1.9.4
cffi==1.16.0 
cryptography==42.0.8 
pycparser==2.22